+++ src/Commands/SystemCommands/NewchatphotoCommand.php
@@ -14,6 +14,8 @@ use Longman\TelegramBot\Commands\SystemCommand;
 
 /**
  * New chat photo command
+ *
+ * @todo Remove due to deprecation!
  */
 class NewchatphotoCommand extends SystemCommand
 {
@@ -42,6 +44,8 @@ class NewchatphotoCommand extends SystemCommand
     {
         //$message = $this->getMessage();
         //$new_chat_photo = $message->getNewChatPhoto();
+
+        trigger_error(__CLASS__ . ' is deprecated and will be removed and handled by ' . GenericmessageCommand::class . ' by default in a future release.', E_USER_DEPRECATED);
 
         return parent::execute();
     }